Создаёт модуль show-gallery.js
